Created a simple push/pull mechanic
Currently, the boxes are incredibly massive so they can't be pushed normally. I made the player movement check for walls before moving so we don't get the player jittering in and out of walls when they run in to them, but this also means that currently you can't push multiple boxes. It should be relatively simple to make that possible. I also plan to make boxes fall out of your grasp when they are not supported by ground.