Made boxes able to move on top of each other
This was accomplished by adding a specific case to ground checking where if a box is on top of another box, it childs itself to the other box. This is undone at the beginning of every ground check unless the box's parent is the player because that means it is being grabbed.