Added manual garbage collection to prevent memory buildup
The weirdness reader has to generate a Texture so it can store a screenshot of the weirdness map to read its color. These textures are only used once but the memory of them was still being stored causing to gradually increasing memory build up. There is now a Destroy call after use so that the texture is not stored.